Revert "Golem pathfinding bug"
This reverts commit 22394c80a1d1b1c25cedb6d1fe5493f5c0986b36.

+++ cPath.cpp
@@ -77,7 +77,7 @@ void cPath::ReMake(int *map,int cxdest,int cydest, int t)
 	}
 }
 
-int cPath::NextStep(int *px,int *py,int *cx,int *cy,float *ix,float *iy, int step_length)
+int cPath::NextStep(int *px,int *py,int *cx,int *cy,int *ix,int *iy, int step_length) // TODO: ix/iy floats
 {
 	int move=CONTINUE;
